old/requirements.txt
git+https://github.com/Rapptz/discord.py
google-api-python-client==1.9.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1